add documentation

+++ lista2/src/main/java/pl/zaisd/Iteration.java
@@ -4,8 +4,8 @@ public class Iteration implements PrimeNumbers {
     @Override
     public boolean isPrime(long n) {
         long max = (long) Math.sqrt(n);
-        for (int i = 2; i <= max; i++) {
-            if(n % i == 0) {
+        for (long i = 2L; i <= max; i++) {
+            if(n % i == 0L) {
                 return false;
             }
         }
